fix: extract initialization hooks
